Split utils.h into StringUtils.h / Log.h
- Most uses of utils.h were just for logging functionality
- String formatting functions are in core/StringUtils.h
- Double-to-string munging functions moved to JsonUtils pending removal
- isqrt etc. were moved to MathUtil
- NOMINMAX is globally defined when compiling under MSVC
